Storage II

+++ app/src/main/java/com/example/shopyproject/product/MainActivity.kt
@@ -18,6 +18,7 @@ import com.example.shopyproject.order.OrderActivity
 import com.firebase.ui.auth.AuthUI
 import com.firebase.ui.auth.ErrorCodes
 import com.firebase.ui.auth.IdpResponse
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.firebase.analytics.FirebaseAnalytics
 import com.google.firebase.analytics.ktx.analytics
 import com.google.firebase.analytics.ktx.logEvent
@@ -26,6 +27,7 @@ import com.google.firebase.firestore.DocumentChange
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ListenerRegistration
 import com.google.firebase.ktx.Firebase
+import com.google.firebase.storage.FirebaseStorage
 
 class MainActivity : AppCompatActivity() , OnProductListener, MainAux {
     private lateinit var firebaseAuth: FirebaseAuth
@@ -148,15 +150,35 @@ class MainActivity : AppCompatActivity() , OnProductListener, MainAux {
         AddDialogFragment().show(supportFragmentManager, AddDialogFragment::class.java.simpleName)    }
 
     override fun onLongClick(product: Product) {
-        val db = FirebaseFirestore.getInstance()
-        val productRef = db.collection(Constants.COLL_PRODUCTS)
-        product.id?.let { id ->
-            productRef.document(id)
-                .delete()
-                .addOnFailureListener{
-                    Toast.makeText(this, "Error al eliminar", Toast.LENGTH_SHORT).show()
+        val user = FirebaseAuth.getInstance().currentUser.uid
+        MaterialAlertDialogBuilder(this)
+            .setTitle(R.string.product_dialog_delete_title)
+            .setMessage(R.string.product_dialog_delete_msg)
+            .setPositiveButton(R.string.product_dialog_delete_confirm) {_,_ ->
+                val db = FirebaseFirestore.getInstance()
+                val productRef = db.collection(Constants.COLL_PRODUCTS)
+                product.id?.let { id ->
+                    FirebaseStorage.getInstance()
+                        .reference
+                        .child(user)
+                        .child(Constants.PATH_PRODUCTS_IMAGES)
+                        .child(id)
+                        .delete()
+                        .addOnSuccessListener {
+                            productRef.document(id)
+                                .delete()
+                                .addOnFailureListener{
+                                    Toast.makeText(this, "Error al eliminar registro", Toast.LENGTH_SHORT).show()
+                                }
+                        }
+                        .addOnFailureListener {
+                            Toast.makeText(this, "Error al eliminar la imagen", Toast.LENGTH_SHORT).show()
+
+                        }
                 }
-        }
+            }
+            .setNegativeButton(R.string.dialog_cancel, null)
+            .show()
     }
 
     private fun configAuth(){
